==> Notebooks/5_Model Evaluation.ipynb <==
# 1. Imports

# from google.colab import drive
# drive.mount('/content/gdrive/', force_remount=True)

# %cd gdrive/MyDrive/BT4222 Project Group/Codes

!pip install keras-tuner
!pip install optuna

import pandas as pd
import numpy as np
import tensorflow as tf
from tensorflow import keras
from keras.models import Sequential
from keras import layers
import tensorflow as tf
from keras.utils.vis_utils import plot_model
from keras_tuner import RandomSearch
from keras.models import save_model, load_model
import random
import warnings
warnings.filterwarnings("ignore")
from keras import backend as K
import optuna
from time import time
from sklearn.metrics import f1_score, accuracy_score
import pickle

# 2. Read in datasets

rdn_index = random.sample(range(110248), 60000)

# word2vec: skip-gram
X_train = pd.read_csv('../Word Embedding/emb_sg_train.csv').iloc[rdn_index]
X_test = pd.read_csv('../Word Embedding/emb_sg_test.csv')

# labels
y_train = pd.read_csv('../Data/y_train.csv')['class'].iloc[rdn_index]
y_test = pd.read_csv('../Data/y_test.csv')['class']

embedding_length = X_train.shape[1]
input_dimension = X_train.shape[0]

print(f'Embedding Length: {embedding_length} \nInput Dimension: {input_dimension}')

# 3. Model Tuning

3a. CNN

def build_model_cnn(hp):

    # create model object
    model = keras.Sequential([
      layers.Input(shape=(embedding_length,1)),
      layers.Conv1D(
          filters=hp.Int('conv_1_filter', min_value=32, max_value=128, step=16), # number of filters
          kernel_size=hp.Choice('conv_1_kernel', values = [3,5,7]), # filter size
          activation='relu',
      ),
      layers.MaxPooling1D(),
      layers.Conv1D(
          filters=hp.Int('conv_2_filter', min_value=32, max_value=64, step=16), # number of filters
          kernel_size=hp.Choice('conv_2_kernel', values = [3,5,7]), # filter size
          activation='relu',
      ),
      layers.GlobalMaxPooling1D(), 
      layers.Dense(
          units=hp.Int('dense_1_units', min_value=32, max_value=256, step=16),
          activation='relu'
      ),
      layers.Dropout(
          hp.Choice('dropout_1', values = [0.1, 0.2, 0.3])
      ),
      layers.Dense(
          units=hp.Int('dense_2_units', min_value=32, max_value=256, step=16),
          activation='relu'
      ),
      layers.Dropout(
          hp.Choice('dropout_2', values = [0.1, 0.2, 0.3, 0.4])
      ),
      # output layer    
      layers.Dense(1, activation='sigmoid', name = 'Output') # output
    ])
    #compilation of model
    model.compile(
        optimizer=keras.optimizers.Adam(hp.Choice('learning_rate', values=[1e-2, 1e-3])),
        loss='binary_crossentropy',
        metrics=['accuracy']
        )
    # return model
    return model

#creating randomsearch object
cnn_tuner = RandomSearch(build_model_cnn, objective='val_accuracy', max_trials = 5)

# search best parameter
cnn_tuner.search(X_train, y_train, epochs=10, validation_data=(X_test, y_test))

cnn = cnn_tuner.get_best_models(num_models=1)[0]

#summary of best model
cnn.summary()

# Fit model
history = cnn.fit(X_train, y_train, epochs=10, validation_data=(X_test, y_test), batch_size = 200)

# Save model
save_model(cnn, '../Models/CNN.h5')

3b. LSTM

lstm_num_neurons = [64, 128]
lstm_dropout = [0.1, 0.2]
lstm_rec_dropout = [0.1, 0.2]
dense_1_dropout = [0.1, 0.2, 0.3]

eval_table = pd.DataFrame()

for lstm_num_neurons_val in lstm_num_neurons:
  for lstm_dropout_val in lstm_dropout:
    for lstm_rec_dropout_val in lstm_rec_dropout:
      for dense_1_dropout_val in dense_1_dropout:

        temp_model = keras.Sequential([
          layers.Input(shape=(X_train.shape[1],1)),
          layers.LSTM(
              lstm_num_neurons_val, 
              dropout=lstm_dropout_val, 
              recurrent_dropout=lstm_rec_dropout_val, 
              return_sequences=True
              ),
          layers.GlobalMaxPooling1D(),
          layers.Dense(256, activation='relu'),
          layers.Dropout(dense_1_dropout_val),
          layers.Dense(128, activation='relu'),
          layers.Dense(1, activation='sigmoid')
        ])

        temp_model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])
        temp_model.fit(X_train, y_train, batch_size=1000, epochs=5)

        y_prob = temp_model.predict(X_test).reshape(-1,1)
        y_pred = list(map(lambda x: 1 if x > 0.5 else 0, y_prob))
        eval_table = eval_table.append({
            'Params': {
                'LSTM_Number of neurons': lstm_num_neurons_val,
                'LSTM_Dropout': lstm_dropout_val,
                'LSTM_Rec_Dropout': lstm_rec_dropout_val,
                'Dense_Dropout': dense_1_dropout_val
            },
            'Accuracy': accuracy_score(y_pred, y_test),
            'Recall': recall_score(y_pred, y_test),
            'Precision': precision_score(y_pred, y_test),
            'F1-Score': f1_score(y_pred, y_test)
        }, ignore_index = True)

# Get the top F1-Score + top Accuracy score
subset = eval_table[eval_table['F1-Score'] == eval_table['F1-Score'].max()]
subset = subset[subset['Accuracy'] == subset['Accuracy'].max()].iloc[0,:]
print(subset)
params = subset['Params']

# Build final model
final_lstm_model = keras.Sequential([
          layers.Input(shape=(X_train.shape[1],1)),
          layers.LSTM(
              params['LSTM_Number of neurons'], 
              dropout=params['LSTM_Dropout'], 
              recurrent_dropout=params['LSTM_Rec_Dropout'], 
              return_sequences=True
              ),
          layers.GlobalMaxPooling1D(),
          layers.Dense(256, activation='relu'),
          layers.Dropout(params['Dense_Dropout']),
          layers.Dense(256, activation='relu'),
          layers.Dense(1, activation='sigmoid')
        ])

final_lstm_model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])
final_lstm_model.fit(X_train, y_train, batch_size=1000, epochs=10)

# save model
# codes to predict
# y_prob = temp_model.predict(X_test).reshape(-1,1)
# y_pred = list(map(lambda x: 1 if x > 0.5 else 0, y_prob))

# Save model
save_model(final_lstm_model, '../Models/LSTM.h5')

3c. XGBoost

Maximize for F1 Score.

***Params***:
- **max_depth**: control max depth of tree, control overfitting
- **learning_rate**: to prevent overfitting
- **gamma**: specifies the minimum loss reduction required to make a node split, makes algorithm conservative
- **reg_lambda**: L2 regularization on weights, makes model conservative
- **subsample**: use random subset of data -> speed up cross validation + prevent overfitting
- **colsample_bytree**: use random subset of features -> speed up cross validation + prevent overfitting

# surpress optuna trial message
optuna.logging.set_verbosity(optuna.logging.WARNING)

import xgboost as xgb

# Run this code to tune our hyperparameters via optuna
def objective(trial):

    params = {
        'objective': 'binary:logistic',
        'max_depth': trial.suggest_int("max_depth", 3, 10), 
        'learning_rate': trial.suggest_float("learning_rate", 0.01, 0.2),
        'gamma': trial.suggest_float("gamma", 0, 1.0),
        'reg_lambda': trial.suggest_float('reg_lambda', 0, 10.0),
        'subsample': trial.suggest_float("subsample", 0.7, 1.0),
        'colsample_bytree': trial.suggest_float("colsample_bytree", 0.5, 1.0),
        'n_estimators': 10                  
    }

    model = xgb.XGBClassifier(**params)

    model.fit(X_train,
              y_train,
              verbose=0,
              early_stopping_rounds=10,
              eval_metric='auc',                # use roc_auc because of balanced data
              eval_set=[(X_test, y_test)])
  
    prediction = model.predict(X_test)
    f1 = f1_score(y_test, prediction)
    return f1

study = optuna.create_study(direction='maximize')
study.optimize(objective, n_trials=100)         # change n_trials as see fit

print("Best trial:")
trial = study.best_trial
print("  F1-Value: ",  trial.value)
print("  Params: ")
for key, value in trial.params.items():
    print(".   {}: {}".format(key, value))

# save model
file_name = "../Models/xgb.pkl"
xgb_model = xgb.XGBClassifier(**trial.params) 
xgb_model.fit(X_train, y_train, verbose=0, early_stopping_rounds=10, eval_metric='auc', eval_set=[(X_test, y_test)]) # use roc_auc because of balanced data
pickle.dump(xgb_model, open(file_name, "wb"))

# load
xgb_model_loaded = pickle.load(open(file_name, "rb"))

3d. LightGBM

import lightgbm as lgb
import warnings
warnings.filterwarnings("ignore")

# Change n_estimators and n_trials as see fit as it affects runtime!!!!

def objective(trial):

    params = {
        'objective': 'binary',
        'boosting': 'gbdt',
        'metric': 'AUC',
        'num_boost_round': 10,                       # change
        'num_leaves': trial.suggest_int("num_leaves", 2, 100),  
        'max_depth': trial.suggest_int("max_depth", 3, 10), 
        'learning_rate': trial.suggest_float("learning_rate", 0.01, 0.2),        
        'feature_fraction': trial.suggest_float('feature_fraction', 0.7, 1.0),
        'bagging_fraction': trial.suggest_float('bagging_fraction', 0.5, 1.0),
        'bagging_freq': trial.suggest_int('bagging_freq', 1, 7),
    }

    
    model = lgb.LGBMClassifier(**params)

    model.fit(X_train,
              y_train,
              verbose=False,
              early_stopping_rounds=10,
              eval_metric='auc',                # use roc_auc because of balanced data
              eval_set=[(X_test, y_test)])
  
    prediction = model.predict(X_test)
    f1 = f1_score(y_test, prediction)
    return f1

study = optuna.create_study(direction='maximize')
study.optimize(objective, n_trials=10)         # change n_trials as see fit

print("Best trial:")
trial = study.best_trial
print("  F1-Value: ",  trial.value)
print("  Params: ")
for key, value in trial.params.items():
    print(".   {}: {}".format(key, value))

# save model
file_name = "../Models/lgb.pkl"
lgb_model = lgb.LGBMClassifier(**trial.params) 
lgb_model.fit(X_train, y_train, verbose=0, early_stopping_rounds=10, eval_metric='auc', eval_set=[(X_test, y_test)]) # use roc_auc because of balanced data
pickle.dump(lgb_model, open(file_name, "wb"))

# load
lgb_model_loaded = pickle.load(open(file_name, "rb"))

3e. SVM

from sklearn.svm import SVC

def objective(trial):

    params = {
        'kernel': trial.suggest_categorical('kernel', ['linear', 'poly', 'rbf', 'sigmoid']),
        'C': trial.suggest_float('C', 0.01, 10),
        'degree': trial.suggest_int('degree', 1, 5, 1),
    }

    model = SVC(**params)

    model.fit(X_train, y_train)
  
    prediction = model.predict(X_test)
    f1 = f1_score(y_test, prediction)
    return f1

study = optuna.create_study(direction='maximize')
study.optimize(objective, n_trials=25)         # change n_trials as see fit

print("Best trial:")
trial = study.best_trial
print("  F1-Value: ",  trial.value)
print("  Params: ")
for key, value in trial.params.items():
    print(".   {}: {}".format(key, value))

# save model
file_name = "../Models/svm.pkl"
svm_model = SVC(**trial.params) 
svm_model.fit(X_train, y_train)
pickle.dump(svm_model, open(file_name, "wb"))

# load
svm_model_loaded = pickle.load(open(file_name, "rb"))

3f. Logistic Regression

from sklearn.linear_model import LogisticRegression

def create_model(trial):

    regularization = trial.suggest_float('C', 0.01, 100)
    model = LogisticRegression(C=regularization, max_iter=150)
    if trial.should_prune():
            raise optuna.TrialPruned()
            
    return model


def objective(trial):
    model = create_model(trial)
    model.fit(X_train, y_train)

    prediction = model.predict(X_test)
    f1 = f1_score(y_test, prediction)
    return f1

study = optuna.create_study(direction='maximize')
study.optimize(objective, n_trials=25)         # change n_trials as see fit

print("Best trial:")
trial = study.best_trial
print("  F1-Value: ",  trial.value)
print("  Params: ")
for key, value in trial.params.items():
    print(".   {}: {}".format(key, value))

# save model
file_name = "../Models/lg.pkl"
lg_model = LogisticRegression(**trial.params) 
lg_model.fit(X_train, y_train)
pickle.dump(lg_model, open(file_name, "wb"))

# load
lg_model_loaded = pickle.load(open(file_name, "rb"))

3g. Random Forest

from sklearn.ensemble import RandomForestClassifier

def objective(trial):

    params = {
        'n_estimators': trial.suggest_int('n_estimators', 50, 200),
        'max_depth': trial.suggest_int('max_depth', 5, 50),
        'min_samples_split': trial.suggest_int('min_samples_split', 2, 20),
        'min_samples_leaf': trial.suggest_int('min_samples_leaf', 2, 20)
    }

    model = RandomForestClassifier(**params)
    
    model.fit(X_train, y_train)
  
    prediction = model.predict(X_test)
    f1 = f1_score(y_test, prediction)
    return f1

study = optuna.create_study(direction='maximize')
study.optimize(objective, n_trials=25)         # change n_trials as see fit

print("Best trial:")
trial = study.best_trial
print("  F1-Value: ",  trial.value)
print("  Params: ")
for key, value in trial.params.items():
    print(".   {}: {}".format(key, value))

# save model
file_name = "../Models/rnd.pkl"
rnd_model = RandomForestClassifier(**trial.params) 
rnd_model.fit(X_train, y_train)
pickle.dump(rnd_model, open(file_name, "wb"))

# load
rnd_model_loaded = pickle.load(open(file_name, "rb"))

# 4. Model Evaluation Results

from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score, roc_auc_score
from keras.models import save_model, load_model

model_results = pd.DataFrame()

# Models that only use embedding as X values
pure_text_models = {
    'LSTM': load_model('../Models/LSTM.h5'),
    'XGBoost': pickle.load(open('../Models/xgb.pkl', "rb")),
    'LightGBM': pickle.load(open('../Models/lgb.pkl', "rb")),
    'SVM': pickle.load(open('../Models/svm.pkl', "rb")),
    'Logistic Reg': pickle.load(open('../Models/lg.pkl', "rb")),
    'Random Forest': pickle.load(open('../Models/rnd.pkl', "rb")),
    'CNN': load_model('../Models/CNN.h5'),
}

for model_name, model in pure_text_models.items():
  y_pred = model.predict(X_test)
  try:
    accuracy = accuracy_score(y_test, y_pred)
    f1 = f1_score(y_test, y_pred, average='weighted')
    precision = precision_score(y_test, y_pred)
    recall = recall_score(y_test, y_pred)
  except ValueError as e:
    y_pred_class = list(map(lambda x: 1 if x > 0.5 else 0, y_pred))
    accuracy = accuracy_score(y_test, y_pred_class)
    f1 = f1_score(y_test, y_pred_class, average='weighted')
    precision = precision_score(y_test, y_pred_class)
    recall = recall_score(y_test, y_pred_class)
  model_results = model_results.append(
      {
       'Model': model_name, 
       'Accuracy': accuracy, 
       'F1-score': f1,
       'Precision': precision,
       'Recall': recall,
      #  'ROC AUC': roc_auc
       }, ignore_index = True)


pd.DataFrame(model_results).to_csv('../Model_Evaluation.csv', index=False)

model_results

# with BERT
pd.read_csv('../Model_Evaluation.csv')
